==> 2. Text operations.ipynb <==
## Performing text operations

import os
import spacy 
from spacy.lang.am import Amharic
from collections import Counter
import pandas as pd

nlp = Amharic()
nlp.max_length = 20000000

data_path = r'C:\Users\user\Documents\Data_science\IR real\Demo_data\Updated_Demo'

books = os.listdir(data_path)

tokens = []

for book in books:
    with open(os.path.join(data_path, book), 'r', encoding='utf-8') as file:
        for line in file:
            line = line.strip()
            doc = nlp(line)
            tokens.extend([token.text for token in doc if not token.is_punct])

#doc = nlp(sentence)

freq_counter = Counter(tokens) #
df = pd.DataFrame({"words" : freq_counter.keys(), "frequency" : freq_counter.values()}).set_index('words')




## Removing numerical values


df = df.reset_index()
df['words'].str.isnumeric().sum()

### This function removes the number that are found in the string

df = df.loc[~df['words'].str.isnumeric()]
df['words'].str.isnumeric().sum()

df.set_index('words', inplace = True)

## Creating frequency and words dataFrame for easier visualization

df.sort_values(by = 'frequency', ascending = False, inplace = True)
df.drop(index = ' ', inplace = True)
df.drop(index = '=', inplace = True)
df.head()

## Creating the Rank

rank = list(range(1, len(df) + 1))
df['Rank']= rank

## Calculating the the normalized constant

df['Constant c'] =(df.frequency / (df.frequency.sum())) * df.Rank
df.head()

## Saving to csv format  

df.to_csv('for_stat.csv')

## These are the total unique number of words


df.shape[0]

## These are the total number of words


df.frequency.sum()

## The top 10 words

df.head(10)

## Saving the index files to csv
* Why we chose these specific numbers could be illustrated using `Statistical analysis.ipynb`


index = df[15:9000]

## The top 5 of words of the index terms

index.head()

df.to_csv('index.csv')
