==> 4. Normalization.ipynb <==
## Importing the extracted data

import pandas as pd 
df = pd.read_csv('for_stat.csv')
df.head()

### Further analsys shows that words with frequencies of 1 are useless so we drop them

df[df['frequency']==1]

df = df[df['frequency'] != 1]

## why we need Normalization

df.loc[df['words'].str.contains('ው')].head()

df.loc[df['words'].str.contains('ዉ')].head()

### As we can see form  the above `ነዉ` and `ነው` are expressing the same word but using different forms to write. So we need to normalize it.

## Creating a dictionary to Normalize the text

dict1 = {
    "ሐ": "ሀ",
    "ሑ": "ሁ",
    "ሒ": "ሂ",
    "ሓ": "ሃ",
    "ሔ": "ሄ",
    "ሕ": "ህ",
    "ሖ": "ሆ",
    "ኀ": "ሀ",
    "ኁ": "ሁ",
    "ኂ": "ሂ",
    "ኃ": "ሃ",
    "ኄ": "ሄ",
    "ኅ": "ህ",
    "ኆ": "ሆ"
}

dict2 = {
    "ሠ": "ሰ",
    "ሡ": "ሱ",
    "ሢ": "ሲ",
    "ሣ": "ሳ",
    "ሤ": "ሴ",
    "ሥ": "ስ",
    "ሦ": "ሶ",
    "ሧ": "ሷ"
}

dict3 = {
    "ዐ": "አ",
    "ዑ": "ኡ",
    "ዒ": "ኢ",
    "ዓ": "ኣ",
    "ዔ": "ኤ",
    "ዕ": "እ",
    "ዖ": "ኦ"
}

dict4 = {
    "ጸ": "ፀ",
    "ጹ": "ፁ",
    "ጺ": "ፂ",
    "ጻ": "ፃ",
    "ጼ": "ፄ",
    "ጽ": "ፅ",
    "ጾ": "ፆ"
}

dict5 = {
    'ዉ' : 'ው'
}

# Merge the dictionaries
merged_dict = {}
merged_dict.update(dict1)
merged_dict.update(dict2)
merged_dict.update(dict3)
merged_dict.update(dict4)
merged_dict.update(dict5)


## Normalization

df.shape

#### Notice how the two words will be the same after normalization

df.loc[df['words']==('ነው')]

df.loc[df['words']==('ነዉ')]

#### We remove other columns. We will only use columns words and frequency because the others will be altered

col = ['words', 'frequency']
df_to_Normalize = df[col].copy()

#### As you can see the number of columns changed from 203030 to 197121 after normalization

def replace_chars(word):
    for char, value in merged_dict.items():
        word = word.replace(char, value)
    return word

df['words'] = df['words'].apply(lambda x: replace_chars(x))


df_Normalize = df.groupby('words')['frequency'].sum().reset_index()
df_Normalize.shape

#### Notice how the frequence is the addition of the two frequences of `ነዉ` and `ነው`

df_Normalize.loc[df_Normalize['words']==('ነው')]

#### Now we re-calculate the rank and constant c

df_Normalize.sort_values(by = 'frequency', ascending = False, inplace = True)

rank = list(range(1, len(df_Normalize) + 1))
df_Normalize['Rank']= rank

df_Normalize['Constant c'] =(df_Normalize.frequency / (df_Normalize.frequency.sum())) * df_Normalize.Rank
df_Normalize.head()

### We will save the normalized data in csv format

df_Normalize.to_csv('Normalized.csv')

index = df_Normalize[15:9000]

### We will save the normalized index in csv format

index.to_csv('Normalized_index.csv')
