==> 6. Inverted Index.ipynb <==
# Indexing 

Amharic_suffix = ['ቻችን','ቻችሁ','ቻቸው','ቻቸውን','ህ','ሽ','ችን','ችሁ','ቸው','ች','ቼ','ቿ', 'ችው', 'ቹ', 'ን']
Amharic_prefix = ['የ', 'በ', 'ለ', 'ከ']
def stemmer(word):
    for suffix in Amharic_suffix:
        word = word.removesuffix(suffix)
    for prefix in Amharic_prefix:
        word = word.removeprefix(prefix)
    return word

## The below code reads the extracted text

import os
import spacy 
from spacy.lang.am import Amharic
from collections import Counter
import pandas as pd

nlp = Amharic()
nlp.max_length = 20000000

data_path = r'C:\Users\user\Documents\Data_science\IR real\Demo_data\Updated_Demo'

books = os.listdir(data_path)


token_file = {}

for book in books:
    with open(os.path.join(data_path, book), 'r', encoding='utf-8') as file:
        for line in file:
            tokens = []
            line = line.strip()
            doc = nlp(line)
            tokens.append([token.text for token in doc if not token.is_punct])
            for token in tokens:
                for word in token:
                    if not word.isnumeric() and not word.isspace():
                        if word not in token_file:
                            token_file[word] = []
                        token_file[word].append(book)

## This is a way to represent the directory path of the books using numbers.

data_path = r'C:\Users\user\Documents\Data_science\IR real\Demo_data\Updated_Demo'

Doc_dict = {}
books = os.listdir(data_path)
index = 0
for book in books:
    Doc_dict[book] = index
    index += 1

## This counts the number of document frequency for each word

words = token_file.keys()
documents = token_file.values()
Doc_index = []
for document in documents:
    indi_Doc = []
    for ind_doc in document:
        index = Doc_dict[ind_doc]
        indi_Doc.append(index)
    Doc_index.append(indi_Doc)
Ifile = pd.DataFrame({"Term" : words, "Doc" : Doc_index})
Ifile[170:175]


### In the below code we will normalize the words. But letter we will show the stemmed/normalized words yields the same result with the unstemmed/un-normalized

Ifile['stemmed'] = Ifile['Term'].apply(lambda x: stemmer(x))
Ifile[170:175]

### By using this simple line of code we can cleary classify count the document frequency and term frequency.

Ifile['CF'] = Ifile['Doc'].apply(len)
Ifile['DF'] = Ifile['Doc'].apply(lambda x: len(set(x)))
Ifile[170:175]

Ifile['Doc'] = Ifile['Doc'].apply(lambda x: set(x))
Ifile[170:175]
