==> 7. VSM_Model.ipynb <==
# Vector space model

### We tokenize and remove punctuation form the extracted text for further processing

import os
import spacy 
from spacy.lang.am import Amharic
from collections import Counter
import pandas as pd

nlp = Amharic()
nlp.max_length = 20000000

data_path = r'C:\Users\user\Documents\Data_science\IR real\Demo_data\Updated_Demo'

books = os.listdir(data_path)


token_file = {}

for book in books:
    with open(os.path.join(data_path, book), 'r', encoding='utf-8') as file:
        for line in file:
            tokens = []
            line = line.strip()
            doc = nlp(line)
            tokens.append([token.text for token in doc if not token.is_punct])
            for token in tokens:
                for word in token:
                    if not word.isnumeric() and not word.isspace():
                        if word not in token_file:
                            token_file[word] = []
                        token_file[word].append(book)

import os
data_path = r'C:\Users\user\Documents\Data_science\IR real\Demo_data\Updated_Demo'

books = os.listdir(data_path)

words = token_file.keys()
documents = token_file.values()

Ifile = pd.DataFrame({"Term" : words, "Doc" : documents})
Ifile[170:175]

Ifile['CF'] = Ifile['Doc'].apply(len)
Ifile['DF'] = Ifile['Doc'].apply(lambda x: len(set(x)))
Ifile[170:175]

### The above code is clearly explained in the `inverted index.ipynb` file

### The below code is used to calcualte the term frequency in each document.

def Tf_counter(docs):
    my_list = {}
    for doc in docs:
        if doc not in my_list:
            my_list[doc] = 1
        else:
            my_list[doc] += 1
    return my_list

Ifile['Doc_freq'] = Ifile['Doc'].apply(lambda x: Tf_counter(x))
Ifile[170:175]

### After calculating the term frequency we will have to change each documents to columns for easier manipulation.

Ifile = pd.concat([Ifile.drop('Doc_freq', axis =1), Ifile['Doc_freq'].apply(pd.Series)], axis = 1)
Ifile[170:175]

Ifile.fillna(0, inplace= True)

### The below code calculates the `IDF`

import numpy as np
def IDF_calc(DF):
    return np.round(np.log10(len(books)/DF), 3)

Ifile['IDF'] = Ifile['DF'].apply(lambda x: IDF_calc(x))
Ifile[170:175]

### Then we calculate the  `tf * idf` so that we know the weight

for book in books:
    Ifile[book] = Ifile[book] * Ifile['IDF']
Ifile[170:175]

import pandas as pd
Ifile = pd.read_csv('IDF_ized.csv')
Ifile[170:172]

Ifile.drop('Unnamed: 0', axis = 1, inplace= True)

## The query processing part

#### This part will calculate the vector length for each book

import numpy as np
book_norm = {}
for book in books:
    book_value = Ifile[book].values
    vector_length = np.round(np.linalg.norm(book_value), 3)
    book_norm[book] = vector_length

#### The below function will prepare the query the same way as inverted index is prepared

import os
from spacy.lang.am import Amharic
from collections import Counter
nlp = Amharic()
nlp.max_length = 20000000
tokens = []

def query(text):
    
    text = text.strip()
    doc = nlp(text)
    tokens.extend([token.text for token in doc if not token.is_punct])

    
    Doc_freq = []
    book_list = []
    trial = {}
    IDF =[]
    query_df = None
    
    for token in tokens:
        word_freq = {}
        for book in books:
            document = os.path.join(data_path, book)
            with open(document, 'r', encoding='utf-8') as file:
                content = file.read()
                word_count= Counter(content.split())
                word_freq[book] = word_count[token]
                
        idf = float(Ifile['IDF'].loc[Ifile['Term'] == token])
        IDF.append(idf)
        
        Doc_freq.append(word_freq)

    query_df = pd.DataFrame({'Term' : tokens, "Doc_freq" : Doc_freq, "IDF" : IDF})
    
    return query_df

#### Let's demonstrate how it works

text = "ባሕር ተሻግሮ ሕይወትን መሸጥ"
query_df = (query(text))
query_df

#### Right here `Q` is the sum of frequency for each word

query_df['Q'] = query_df['Doc_freq'].apply(lambda x: np.sum(list(x.values())))

#### Right here `Q` is the sum of frequency for each word multiplied by the `IDF` therefore it represents the weight

query_df['Q'] = query_df['Q'] * query_df['IDF']
query_df.drop(['Doc_freq', 'IDF'],axis =1, inplace=True)
query_df

##### This will list only the words that are present in the query and inverted index file. It will save it to `Ifile`

Ifile = Ifile.loc[Ifile['Term'].isin(tokens)]

Ifile = Ifile[['Term'] + books]

##### We will merge the query dataframe and the words from the inverted index file to `QI_df`

QI_df = query_df.merge(Ifile, on='Term')

##### Now by multiplying `QI_df[book]` with `QI_df['Q']` we prepare the columns for the dot product. Therefore when we add the columns we will have the dot product.

for book in books:
    QI_df[book] = QI_df[book] * QI_df['Q']

QI_df

#### Vector length of  `Q`

query_value = QI_df[book].values
vector_length = np.round(np.linalg.norm(query_value), 3)


##### We compute similarity score and sort them in descending value

vsm_result = {}
for book in books:
    result = QI_df[book].sum()
    result = result / (vector_length * book_norm[book])
    vsm_result[book] = result

vsm_result = pd.Series(vsm_result, name = 'result')
vsm_result = pd.DataFrame(vsm_result)
vsm_result.sort_values(by='result', ascending= False, inplace=True)

vsm_result
